+++ packages/popmotion-pose/src/dom/events.ts
@@ -2,35 +2,87 @@ import { listen, ColdSubscription } from 'popmotion';
 import { ActiveActions } from 'pose-core/lib/types';
 import { DomPopmotionConfig } from '../types';
 
+/**
+ * TODO: Currently we add a new listener for every event,
+ * it'd be better to add one listener for every pointer event and use
+ * event delegation
+ */
+
 type Setter = (next: string) => Promise<any>;
 
-const makeDraggable = (
+type PointerEventsConfig = {
+  startEvents: string;
+  endEvents: string;
+  startPose: string | string[];
+  endPose: string | string[];
+  startCallback: string;
+  endCallback: string;
+  startListener: string;
+  endListener: string;
+  useDocumentToEnd?: boolean;
+};
+
+const makePointerEvent = ({
+  startEvents,
+  endEvents,
+  startPose,
+  endPose,
+  startCallback,
+  endCallback,
+  startListener,
+  endListener,
+  useDocumentToEnd
+}: PointerEventsConfig) => (
   element: Element,
   activeActions: ActiveActions<ColdSubscription>,
   setPose: Setter,
-  { onDragStart, onDragEnd }: DomPopmotionConfig
+  config: DomPopmotionConfig
 ) => {
-  const dragStartListener = listen(element, 'mousedown touchstart').start(
+  const eventStartListener = listen(element, startEvents).start(
     (startEvent: MouseEvent | TouchEvent) => {
       startEvent.preventDefault();
-      setPose('dragging');
+      Array.isArray(startPose) ? startPose.map(setPose) : setPose(startPose);
 
-      if (onDragStart) onDragStart(startEvent);
+      if (config[startCallback]) config[startCallback](startEvent);
 
-      const dragEndListener = listen(document, 'mouseup touchend').start(
-        (endEvent: MouseEvent | TouchEvent) => {
-          activeActions.get('dragEndListener').stop();
-          setPose('dragEnd');
-          if (onDragEnd) onDragEnd(endEvent);
-        }
-      );
+      const eventEndListener = listen(
+        useDocumentToEnd ? document : element,
+        endEvents
+      ).start((endEvent: MouseEvent | TouchEvent) => {
+        activeActions.get(endListener).stop();
+        Array.isArray(endPose) ? endPose.map(setPose) : setPose(endPose);
+        if (config[endCallback]) config[endCallback](endEvent);
+      });
 
-      activeActions.set('dragEndListener', dragEndListener);
+      activeActions.set(endListener, eventEndListener);
     }
   );
 
-  activeActions.set('dragStartListener', dragStartListener);
+  activeActions.set(startListener, eventStartListener);
 };
+
+const makeDraggable = makePointerEvent({
+  startEvents: 'mousedown touchstart',
+  endEvents: 'mouseup touchend',
+  startPose: 'dragging',
+  endPose: 'dragEnd',
+  startCallback: 'onDragStart',
+  endCallback: 'onDragEnd',
+  startListener: 'dragStartListener',
+  endListener: 'dragEndListener',
+  useDocumentToEnd: true
+});
+
+const makeHoverable = makePointerEvent({
+  startEvents: 'mouseenter',
+  endEvents: 'mouseleave',
+  startPose: 'hovering',
+  endPose: ['default', 'hoverEnd'],
+  startCallback: 'onHoverStart',
+  endCallback: 'onHoverEnd',
+  startListener: 'hoverStartListener',
+  endListener: 'hoverEndListener'
+});
 
 export default (
   element: Element,
@@ -39,4 +91,5 @@ export default (
   { props }: DomPopmotionConfig
 ) => {
   if (props.draggable) makeDraggable(element, activeActions, setPose, props);
+  if (props.hoverable) makeHoverable(element, activeActions, setPose, props);
 };