+++ packages/popmotion-pose/src/types.ts
@@ -38,7 +38,10 @@ export interface DomPopmotionPoser
 export type DomPopmotionConfig = {
   onDragStart?: (e: MouseEvent | TouchEvent) => any;
   onDragEnd?: (e: MouseEvent | TouchEvent) => any;
+  onHoverStart?: (e: MouseEvent | TouchEvent) => any;
+  onHoverEnd?: (e: MouseEvent | TouchEvent) => any;
   draggable?: boolean;
+  hoverable?: boolean;
   dragBounds?: BoundingBox;
   props?: { [key: string]: any };
   [key: string]: any;