+++ packages/pose-core/src/_tests/index.test.ts
@@ -62,7 +62,8 @@ const testPose = poseFactory<Value, Action, Subscription, PoserAPI>({
   defaultTransitions: new Map([['default', ({ to }) => mockActionInverse(to)]]),
   transformPose: pose => pose,
   readValueFromSource: () => 0,
-  extendAPI: api => api
+  extendAPI: api => api,
+  posePriority: ['draggable', 'hoverable']
 });
 
 const testPoser = testPose({
@@ -214,3 +215,60 @@ test('resolves custom transitions correctly', () =>
       expect(testPoser.get().x).toBe(45)
     })
 );
+
+test('applies correct initial values', () => {
+  const testDefaultsPoser = testPose({
+    default: { x: 100 },
+    a: { x: 50 }
+  });
+
+  expect(testDefaultsPoser.get('x')).toBe(100);
+
+  const testDefaultsOverridePoser = testPose({
+    default: { x: 100 },
+    a: { x: 50 },
+    initialPose: 'a'
+  });
+
+  expect(testDefaultsOverridePoser.get('x')).toBe(50);
+});
+
+test('correctly falls back to previous pose', () => {
+  const fallback = testPose({
+    default: { x: 0 },
+    a: { x: 50 }
+  });
+  expect(fallback.get('x')).toBe(0);
+  return fallback.set('a').then(() => {
+    expect(fallback.get('x')).toBe(-50);
+    return fallback.set('a', false);
+  })
+  .then(() => {
+    expect(fallback.get('x')).toBe(0);
+  });
+});
+
+test('correctly applies poses in priority order', () => {
+  const fallback = testPose({
+    default: { x: 0 },
+    draggable: { x: 10 },
+    hoverable: { x: 20 }
+  });
+  expect(fallback.get('x')).toBe(0);
+  return fallback.set('hoverable').then(() => {
+    expect(fallback.get('x')).toBe(-20);
+    return fallback.set('draggable');
+  })
+  .then(() => {
+    expect(fallback.get('x')).toBe(-10);
+    return fallback.set('hoverable', false)
+  }).then(() => {
+    expect(fallback.get('x')).toBe(-10)
+    return fallback.set('hoverable')
+  }).then(() => {
+    expect(fallback.get('x')).toBe(-10)
+    return fallback.set('draggable', false)
+  }).then(() => {
+    expect(fallback.get('x')).toBe(-20)
+  })
+});