+++ packages/popmotion-pose/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Popmotion Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.2.0] 2018-08-18
+
+### Added
+
+- `hoverable` config option.
+
 ## [2.1.2] 2018-08-13
 
 ### Fixed