+++ packages/pose-core/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Pose Core adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.8.1] 2018-08-18
+
+### Added
+
+- Exporting `TransitionMap` type.
+
 ## [0.8.0] 2018-06-14
 
 ### Added